Fixed folders and some appRelease reference for appDeploy sharedResource.
