+++ src/react-apps/applications/service-development/src/sharedResources/appDeploy/appDeployDispatcher.ts
@@ -1,24 +1,25 @@
 import { Action, ActionCreatorsMapObject, bindActionCreators } from 'redux';
 import { store } from '../../store';
-import * as AppDeploymentActions from './appDeployActions';
+import * as CreateAppDeploymentActions from './create/createAppDeploymentActions';
+import * as GetAppDeploymentActions from './get/getAppDeploymentActions';
 import { IDeployment } from './types';
 
 export interface IAppReleaseActionDispatcher extends ActionCreatorsMapObject {
   getDeployments: () => Action;
-  getDeploymentsFulfilled: (releases: IDeployment[]) => AppDeploymentActions.IGetDeploymentsFulfilled;
-  getDeploymentsRejected: (error: Error) => AppDeploymentActions.IGetDeploymentsRejected;
-  createDeployment: (tagName: string, envName: string) => AppDeploymentActions.ICreateDeployment;
-  createDeploymentFulfilled: (id: string) => AppDeploymentActions.ICreateDeploymentFulfilled;
-  createDeploymentRejected: (error: Error) => AppDeploymentActions.ICreateDeploymentRejected;
+  getDeploymentsFulfilled: (releases: IDeployment[]) => GetAppDeploymentActions.IGetDeploymentsFulfilled;
+  getDeploymentsRejected: (error: Error) => GetAppDeploymentActions.IGetDeploymentsRejected;
+  createDeployment: (tagName: string, envName: string) => CreateAppDeploymentActions.ICreateDeployment;
+  createDeploymentFulfilled: (id: string) => CreateAppDeploymentActions.ICreateDeploymentFulfilled;
+  createDeploymentRejected: (error: Error) => CreateAppDeploymentActions.ICreateDeploymentRejected;
 }
 
 const actions: IAppReleaseActionDispatcher = {
-  getDeployments: AppDeploymentActions.getDeployments,
-  getDeploymentsFulfilled: AppDeploymentActions.getDeploymentsFulfilled,
-  getDeploymentsRejected: AppDeploymentActions.getDeploymentsRejected,
-  createDeployment: AppDeploymentActions.createRelease,
-  createDeploymentFulfilled: AppDeploymentActions.createReleaseFulfilled,
-  createDeploymentRejected: AppDeploymentActions.createReleaseRejected,
+  getDeployments: GetAppDeploymentActions.getDeployments,
+  getDeploymentsFulfilled: GetAppDeploymentActions.getDeploymentsFulfilled,
+  getDeploymentsRejected: GetAppDeploymentActions.getDeploymentsRejected,
+  createDeployment: CreateAppDeploymentActions.createDeployment,
+  createDeploymentFulfilled: CreateAppDeploymentActions.createDeploymentFulfilled,
+  createDeploymentRejected: CreateAppDeploymentActions.createDeploymentRejected,
 };
 
 const AppReleaseActionDispatcher: IAppReleaseActionDispatcher = bindActionCreators<
